rename nul folder to null
