+++ src/attributes/index.ts
@@ -4,7 +4,7 @@ import { binary } from './binary/index.js'
 import { boolean } from './boolean/index.js'
 import { list } from './list/index.js'
 import { map } from './map/index.js'
-import { nul } from './nul/index.js'
+import { nul } from './null/index.js'
 import { number } from './number/index.js'
 import { record } from './record/index.js'
 import { set } from './set/index.js'
@@ -12,7 +12,7 @@ import { string } from './string/index.js'
 
 export * from './any/index.js'
 export * from './primitive/index.js'
-export * from './nul/index.js'
+export * from './null/index.js'
 export * from './boolean/index.js'
 export * from './number/index.js'
 export * from './string/index.js'
