+++ src/attributes/primitive/typer.unit.test.ts
@@ -7,7 +7,7 @@ import { binary } from '../binary/index.js'
 import { boolean } from '../boolean/index.js'
 import { $state, $type } from '../constants/attributeOptions.js'
 import type { Always, AtLeastOnce, Never } from '../constants/index.js'
-import { nul } from '../nul/typer.js'
+import { nul } from '../null/typer.js'
 import { number } from '../number/index.js'
 import { string } from '../string/index.js'
 import type { Validator } from '../types/validator.js'
