+++ src/schema/actions/fromJSON/attribute/primitive.ts
@@ -1,6 +1,6 @@
 import { binary } from '~/attributes/binary/index.js'
 import { boolean } from '~/attributes/boolean/index.js'
-import { nul } from '~/attributes/nul/index.js'
+import { nul } from '~/attributes/null/index.js'
 import { number } from '~/attributes/number/index.js'
 import type { $PrimitiveAttributeNestedState } from '~/attributes/primitive/index.js'
 import { string } from '~/attributes/string/index.js'
