+++ src/schema/actions/jsonize/jsonize.unit.test.ts
@@ -6,7 +6,7 @@ import { binary } from '~/attributes/binary/index.js'
 import { boolean } from '~/attributes/boolean/index.js'
 import { list } from '~/attributes/list/index.js'
 import { map } from '~/attributes/map/index.js'
-import { nul } from '~/attributes/nul/index.js'
+import { nul } from '~/attributes/null/index.js'
 import { number } from '~/attributes/number/index.js'
 import { record } from '~/attributes/record/index.js'
 import { set } from '~/attributes/set/index.js'
